Add helper methods for GET/POST/PUT/DELETE
These are available for use in the situation that Nexmo add new API
endpoints, but the customer cannot upgrade their library. By reusing
the existing `Client::send` method, we add authentication etc for free.

The desired way to implement this would be for consumers to create
their own `Zend\Diactoros\Request` (or anything that is
`Psr\Http\Message\RequestInterface` compatible) objects to pass to
`Client::send`. This option is still available - these new methods are
for those that want a quick way to get moving.