Download recordings using $client->get()
As there's no way to fetch a list of recordings for a given call
yet, we can't add `::downloadRecording()` as a method on a call
object.

This PR adds the ability to download recordings using the generic
`client::get()` method by detecting if the request is to a
`/v1/file` endpoint and using Bearer auth if so.

Resolves #44