Support for purchasing a number
At the moment we can't distinguish why a purchase has failed, so we
just pass the server error back to the user.

When a number is successfully purchased, there's no useful information
in the response. We could perform a `GET` on the number and return an
object but it feels a little heavy handed.

For now, the method doesn't return anything, and throws exceptions on
any errors. If the method call succeeds, we purchased the number
successfully

Resolves #55