Update
+ Added PHP 8.0 to workflow
+ Optimized workflow
+ Updated composer.json to install PHPUnit 9. when needed

Please note, that all tests using prophecy will throw a deprecation warning, when ran under PHPUnit 9.
To get rid of these, phpspec/prophecy-phpunit needs to be installed. However, the minimum PHP version
for this package is 7.3, thus we have to live with those warnings for now.

If it gets too annoying, remove PHP 8 from the workflow and re-implement it, if the minimum version
for this SDK has been bumped to 7.3.